src/main/java/com/nisum/userapi/serviceImpl/JwtUserDetailsService.java
package com.nisum.userapi.serviceImpl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.ArrayList;

@Service
public class JwtUserDetailsService implements UserDetailsService {
    private final String defaultUsername;
    private final String defaultPassword;

    @Autowired
    public JwtUserDetailsService(@Value("${jwt.username}") String defaultUsername, @Value("${jwt.password}") String defaultPassword) {
        this.defaultUsername = defaultUsername;
        this.defaultPassword = defaultPassword;
    }

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        if (defaultUsername.equals(username)) {
            return new User(defaultUsername, defaultPassword, new ArrayList<>());
        } else {
            throw new UsernameNotFoundException("Usuario no encontrado: " + username);
        }
    }
}
